Validacion 3 intentos Saul

+++ database/migrations/2018_10_27_171556_crate_incidencias_table.php
@@ -16,8 +16,8 @@ class CrateIncidenciasTable extends Migration
         Schema::create('incidencias', function (Blueprint $table) {
 
             $table->increments('id');
-            $table->float('latitud');
-            $table->float('longitud');
+            $table->double('latitud');
+            $table->double('longitud');
             $table->date('fecha');
             $table->string('comentario')
                   ->nullable();
